packages/pipeline/requirements-runtime.txt
# Runtime-only dependencies for Sunday Album.
# This file is bundled inside the .app and used by setup-runtime.sh on first launch
# to create the production Python venv at ~/Library/Application Support/SundayAlbum/venv/.
#
# Dev-only tools (pytest, mypy, ruff, matplotlib, torch, diffusers) are intentionally excluded.

# Core image processing
opencv-python-headless>=4.10.0.84
numpy>=1.26.4,<3
scipy>=1.13.1
scikit-image>=0.24.0
Pillow>=10.4.0

# Format support
pillow-heif>=0.18.0
rawpy>=0.21.0

# CLI and utilities
click>=8.1.0
python-dotenv>=1.0.0

# AI APIs
anthropic>=0.49.0
openai>=1.47.1
